+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 1.0.0-beta
 
 - Migrated to sound null safety
-- Added SharedPref class as a shortcut to PrefItem(SharedPrefsStorage(), ...)
+- Added PrefNotifier class as a shortcut to PrefItem(SharedPrefsStorage(), ...)
 - Multiple minor improvements 
 
 # 0.0.7